Extract base class from SnapshotsTable
